requirements.txt
streamlit>=1.28.0
pymongo>=4.5.0
pillow>=10.0.0
google-generativeai>=0.3.0
bcrypt>=4.0.0
pandas>=2.0.0
altair>=5.0.0
python-dotenv>=1.0.0